# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "API для работы с Task manager",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/id-cards": {
            "get": {
                "tags": [
                    "IdCards"
                ],
                "summary": "Получить список всех ID-карт",
                "operationId": "8b8730072c5b0d1d8ec30579e80ca7b3",
                "responses": {
                    "200": {
                        "description": "Список всех ID-карт успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID ID-карты",
                                                "type": "integer"
                                            },
                                            "user_id": {
                                                "description": "ID пользователя, которому принадлежит ID-карта",
                                                "type": "integer"
                                            },
                                            "card_number": {
                                                "description": "Номер ID-карты",
                                                "type": "string"
                                            },
                                            "issued_at": {
                                                "description": "Дата выдачи ID-карты",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "expires_at": {
                                                "description": "Дата окончания срока действия ID-карты",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "IdCards"
                ],
                "summary": "Создать новую ID-карту",
                "operationId": "4a0f59af5226f5c4adc55b23cd645bd0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_id": {
                                        "description": "ID роли",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID пользователя, которому принадлежит ID-карта",
                                        "type": "integer"
                                    },
                                    "condition_id": {
                                        "description": "ID состояния (необязательно)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "ID-карта успешно создана",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID ID-карты",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID пользователя",
                                            "type": "integer"
                                        },
                                        "card_number": {
                                            "description": "Номер ID-карты",
                                            "type": "string"
                                        },
                                        "issued_at": {
                                            "description": "Дата выдачи ID-карты",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "expires_at": {
                                            "description": "Дата окончания срока действия ID-карты",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/id-cards/{id}": {
            "get": {
                "tags": [
                    "IdCards"
                ],
                "summary": "Получить данные о конкретной ID-карте",
                "operationId": "2a330ba1fa743b1ae183aeffd87fe177",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные ID-карты успешно получены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID ID-карты",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID пользователя",
                                            "type": "integer"
                                        },
                                        "card_number": {
                                            "description": "Номер ID-карты",
                                            "type": "string"
                                        },
                                        "issued_at": {
                                            "description": "Дата выдачи ID-карты",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "expires_at": {
                                            "description": "Дата окончания срока действия ID-карты",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "ID-карта не найдена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "IdCards"
                ],
                "summary": "Обновить данные ID-карты",
                "operationId": "0319d4dd908776ae98b868197cd3b17f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "role_id": {
                                        "description": "ID роли",
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "description": "ID пользователя",
                                        "type": "integer"
                                    },
                                    "condition_id": {
                                        "description": "ID состояния (необязательно)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ID-карта успешно обновлена",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID ID-карты",
                                            "type": "integer"
                                        },
                                        "user_id": {
                                            "description": "ID пользователя",
                                            "type": "integer"
                                        },
                                        "card_number": {
                                            "description": "Номер ID-карты",
                                            "type": "string"
                                        },
                                        "issued_at": {
                                            "description": "Дата выдачи ID-карты",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "expires_at": {
                                            "description": "Дата окончания срока действия ID-карты",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "IdCards"
                ],
                "summary": "Удалить ID-карту",
                "operationId": "8503226ec1c5b9e3a5d87d5df5ad555e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID-карта успешно удалена"
                    }
                }
            }
        },
...
    },
...
}